added runge-kutta
method to solve ordinary differential equation (ODE). for lane-emden equation context, runge-kutta was used for finding numeric solution of lane emden equation, which converted to two ODE. in this code, we can find xi value while theta = 0, or xi1. xi1 is a star surface boundary, which will later be used to determine the physical radius of the star.